Fix type conversion warnings in tests.

+++ tests/check-tsarray_append.c
@@ -53,10 +53,10 @@ END_TEST
  */
 START_TEST(test_append_many)
 {
-    const long start = -1010;
-    const long stop = 32010;
-    const unsigned long expected_len = (unsigned long)(stop-start);
-    unsigned int i;
+    const int start = -1010;
+    const int stop = 32010;
+    const int expected_len = stop - start;
+    int i;
 
     append_seq_checked(a1, start, stop);
